# Remove temporary buffer from AddConvexPolyFilled