@ocornut 2018-05-13
Hello,

Thanks for submitting this.
What are you trying to solve? Assuming it is meant to be a performance improvement: Do you have a process and measurements to share comparing the speed of this version and the current one? I typically compare performance on both fully optimized and debug builds.

As a crude performance test, I have bound the two variants of the code on the CTRL key (adding `if (ImGui::GetIO().KeyCtrl) {}` in that function to select the algorithm) and modified the ShowExampleAppCustomRendering() to keep only the filled primitives and display everything 200 times, then measured average time over 6 seconds:

200 iterations of rendering all the filled shapes:

Win32 Release Optimized:
Before: 1.48 ~ 1.50 ms
After: 1.66 ~ 1.69 ms (slower)

Win64 Release Optimized:
Before: 1.21 ~ 1.24 ms
After: 1.35 ~ 1.38 ms (slower)

Win32 Debug Non-Optimized
Before: 7.27 ~ 7.31 ms
After: 7.70 ~ 7.74 ms (slower)

etc.
So this would need to be reworked because right now unless I messed up my tests this is slower every time.

- avoid the modulo in the loop like the original code did
- for the Debug perf, avoiding calling `ImVec2()` constructors would probably also be beneficial
- the test for the last point should be moved outside the loop (the `[2] [4] [5]` operators would need to be offset probably by `points_count >= 2 ? -9 : -6` (untested)
- etc.

Also smaller/other things that would need be fixed later:
- use imgui coding style (braces, spaces)
- squash the dozens commits into one

Thanks,
Omar