@ocornut 2018-05-13
About this:

> When i have it in my head: could it also auto detect winding order?
> Currently it assumes clockwise winding, which is ok, but isn't documented anywhere.

I have added a comment about anti-aliased filling requiring a path in clockwise order.
I don't think we can detect clockwiseness without a perf hit. We could eventually provide different primitives to either fill a CCW path or optional detect and apply AA properly in all situation but I don't think there is a need for it now.