@G-glop 2018-05-14
Removing alloca is mostly for compatibility. (it's only other use is in AddPolyline, which can be modified in the same way) 

Performance shouldn't have changed. Doing register-register ops ins't a win when pipelining hides the latency of reading memory.

I made a benchmark of processing a rounded rectangle. (not yet pushed) And MSVC's sampling profiler shows 10% of the whole run time is spent on `if(scale < 100.0f)`, but only in some configurations, otherwise performance is identical.

But i'm only measuring the pure run time of the function itself, maybe the fact that the fill triangles are interleaved with the outline degrades performance? If so, it's easy to de-interleave them, it also may be worthwhile to also use a different algorithm  that doesn't produce long, thin triangles.