@ocornut 2018-05-14
> Removing alloca is mostly for compatibility. (it's only other use is in AddPolyline, which can be modified in the same way)

Which compatibility issue do you have with `alloca` ?
I don't mind changing the rendering function especially if we get a small boost or a simplification, but if it's for compatibility we need to also state what the compatibility issue is.

> But i'm only measuring the pure run time of the function itself, maybe the fact that the fill triangles are interleaved with the outline degrades performance? If so, it's easy to de-interleave them, 

I guess it's more reasonable and representative of real use to keep them interleaved. Regardless, a simple manufactured/looping test would probably fit in the instruction cache anyway even with multiple functions running.

> it also may be worthwhile to also use a different algorithm that doesn't produce long, thin triangles.

What's the problem with long thin triangles? 

Mildly unrelated, but if you fancy looking at it, the non-AA thick strokes are broken (see #288, one of the oldest PR alive, it was submitted just at the time we switched to AA primitives).