@G-glop 2018-05-29
Sorry for taking so long, i got sidetracked trying (and failing) to install vTune, to see what's really slowing it down.

But reading the disassembly i think i know what's going on. In the old version the normal computing loop got unrolled to call `sqrtf ` 5 times each iteration (nevermind that `sqrtf ` is a wrapper for AVX sqrtf instruction which computes 8 sqrtfs at once). Meanwhile my version has a critical path of computing the normals sequentially.

Without doing explicit SIMD (intrinsics/ispc) i don't think i can get a (reliable) performance boost. Maybe a compromise with a small fixed size buffer will yield the best result.

About the compatibility issue; `alloca` itself is fine (tho someone will be surprised when using a small stack and drawing complex polygons), the issue reolves around the implied `__chkstk_ms`. Windows only reserves one page above the stack to commit in when the stack grows. This all works fine untilyou have stack frames larger than 4kB, then you could access memory beyond the reserved page and cause a seqfault, so `chkstk` is inserted to sequentially touch the pages.

 MinGW has (had?) a problem with properly handling it, causing link errors.

And lastly the short version why long thin traingles aren't so great:

GPU's first determine which 8x8 tiles a triangle overlaps, in those tiles it applies the same process but for  2x2 groups of pixels, if the traingle is overlapping a group then the shader is executed for all the pixels in that group and pixels outside the triangle are discarded.

Additinaly mobile/newer generation GPUs use tile renderers, which add another level with even larger tiles designed to fit in cache.

[long source](https://fgiesen.wordpress.com/2011/07/06/a-trip-through-the-graphics-pipeline-2011-part-6/)

So long thin triangles hit a lot of tiles and groups, but contribute little pixels to the output.