@G-glop 2018-05-30
Ok, i think i will leave it at this. Using the output as a buffer gets slightly worse performance when it's the only thing in cache and slightly better performace when there's cache pressure.